Go Module. timestamps moved to file. Instructions bin
