+++ deprecated/go/go.txt
@@ -98,3 +98,6 @@ projectId := rawurl[len(url)-1:]
 
 
 // "Host" header provided via sdk in python/event.py but in python/proxy.py (Flask). "Host" not required by Sentry.io
+
+
+4. `go get github.com/google/uuid github.com/mattn/go-sqlite3 github.com/joho/godotenv github.com/shopspring/decimal`
