+++ go.mod
@@ -1,0 +1,12 @@
+module undertaker
+
+go 1.12
+
+require (
+	github.com/buger/goreplay v1.0.0
+	github.com/buger/jsonparser v1.0.0
+	github.com/google/uuid v1.1.1
+	github.com/joho/godotenv v1.3.0
+	github.com/mattn/go-sqlite3 v2.0.3+incompatible
+	github.com/shopspring/decimal v1.2.0
+)
